requirements.txt
pycryptodomex>=3.20.0
loguru>=0.7.2